Make this match the agent v3 code
https://github.com/buildkite/agent/blob/master/clicommand/bootstrap.go#L217-L221

I'm not sure this is the correct way around to fix this, but at the moment, the docs and code seem to be at odds.

I would find it clearer if the variable indicated positively which way around the boolean should be interpreted. So `BUILDKITE_WILL_FETCH_GIT_SUBMODULES` or similar. But, naming is hard :)