consistently wrap slack lines with quotes
Earlier in this page we warn that it's best to wrap the value of these
lines in quotes.

Strictly speaking the quotes are only required for values that start
with a hash (#), but I think it's worth follow our own advice in all the
examples. Just to be clear that it's the recommended approach.