log 404s to stdout
This app is using lograge [1] to log requests one-per line. Like this:

    method=GET path=/docs/tutorials/getting-started format=html controller=PagesController action=show status=200 duration=1771.03 view=44.04

It renders some 404s (like when ActiveRecord::NotFound is raised), but
it's unable to log requests for unmatched routes. The workaround
described in the README [2] is to add a catch-all route at the lowest
priority and render the 404 page with a controller action.

With this change, we got 404 logs:

    method=GET path=/docs-fargate/tutorials/getting-started format=html controller=ApplicationController action=route_not_found status=404 duration=4.73 view=4.52

This will be helpful for debugging the new deployment on fargate.

[1] https://rubygems.org/gems/lograge
[2] https://github.com/roidrage/lograge#handle-actioncontrollerroutingerror