Configure the app to serve pages at /docs/* and /docs-fargate/*
We're preparing to serve https://buildkite.com/docs with this app
deployed to fargate.

During the preparation we'd like to serve https://buildkite.com/docs
from the current location, and https://buildkite.com/docs-fargate from
this app on fargate.

Once we're happy that the rails app on fargate is working correctly, we
can change the ALB to route https://buildkite.com/docs to it.

This PR attempts to make the app capable of working under both route
prefixes, so there's no code changes required to start serving
production traffic (and we can then remove support for /docs-fargate) at
our leisure.