expand sidebar sections when the URL prefix is /docs-fargate
While we're in the process of migrating docs to be hosted on fargate,
we're routing https://buildkite.com/docs-fargate/* to the fargate
containers.

This change allows the sidebar sections to unroll when the fargate
prefix is used.

This is a bit messy for a view, but it's temporary so..... hopefully
it's palatable?