update some routes to work with /docs and /docs-fargate URL prefix
While we're in the process of migrating docs to be hosted on fargate,
we're routing https://buildkite.com/docs-fargate/* to the fargate
containers.

Some sections of the documentation have no index page and we use a route
to redirect to the first sub-page. This change ensures those redirects
work with both /docs/* and /docs-fargate/*