Add a deploy step if the build is green
A step towards continuously deploying the docs site to fargate.

I assume this might take a few cycles to get right, but hopefully this is
a good start. We're still feeling out way with fargate, so I've based
this on the way we deploy site.

There seems to be rough consensus that we want to store the basic task
definition inside the code repo (or at least, the key parameters that
will vary over time. Like CPU, memory, and image).

It's not yet clear what our ideal tooling around that will look like. I
assume we'll eventually want something more capable than a shell script.
However, until we settle on what that future looks like I've emulated
the way we deploy site to ECS. At least when we do decide on more
capable tooling, we can migrate docs and site in the same way.