revert back to deploying with a custom script
In PR #678, we decided to try deploying to ecs/fargate with the
ecs-deploy plugin.

That turned out to be tricky. After a few rounds of debugging, it's
exiting with the following error:

    --- Registering new task definition for docs
    Registered docs:16
    🚨 Error: The command exited with status 1

I'd like to drop back to a simple script and see that working, then I might try
the plugin again.