backport algolia search support from buildkite.com
A few weeks ago, we added search to https://buildkite.com/docs, using
algolia [1].

We're in the process of extracting the docs content to be hosted as a
dedicated, self-contained rails app that's deployed to fargate. Before
that migration can be made live for users, we first need to backport the
new search functionality to this app so it'll remain available for
users.

Most of this is copied as-is from its original home, and it seems to
work.

I've opted to store the algolia config values in SSM for consistency.
On the one hand their public values that can be seen in the page source,
but on the other hand they're config and it feel weird to just store
them in a view. I'm happy to reconsider if anyone feels strongly.

I also had to hack the CSP settings so the algolia JS would load. I
haven't done much work with CSP so I assume I've opened a big security
hole and would love someone to take a look at it.

[1] https://www.algolia.com/