add a file extension to this bazel image name
leaving the file extension off works fine in development, but we're
getting an exception in the preview app:

    Sprockets::Rails::Helper::AssetNotFound (The asset "tutorials/bazel/add-to-bazel" is not present in the asset pipeline.)

I'm not familiar with our customer image URL tag helper (Page#image),
but I noticed we typically pass image extensions to it. Maybe we'll get
lucky and this will fix the exception?