Proof of concept: parse step attributes out of existing Markdown content
This replaces the manually-created, multi-file step definitions with parsing existing information out of our Markdown pages. This is to avoid requiring that all docs updates take place in multiple places, and the possible future where a page is defined by multiple files.

Instead, the existing table structures we use in the docs are parsed for attributes, and the title and description are parsed out of the document itself. It's entirely possible to customise this processing further to suit our needs better, but as a first-pass proof of concept I feel like this demonstrates the approach well.

One other unsolved problem is that we can't use fenced code blocks within tables currently. We don't use those yet, but the new reference documentation did; if we want support for it within the descriptions we will need to move both the docs and the reference renderer to a CommonMark-compliant parser, as redcarpet is not compliant, and the parser used by markdown-to-jsx is also not compliant, and they each have differing behaviour. We were going to hit this issue with this project at some point, but I'd rather it not be post-cycle.