Improve link styles
The links in the app were imported from Buildkite app in 2019 and
haven't been updated since.

We had an issue with `code` that are links or part of link text looking
odd or not consistently like a link. This commit re-styles links across
the whole site (!) but makes them consistent, including those with
`code`.