AzureAD SCIM: simplify instructions
We've had feedback from a customer who has enabled SCIM with AzureAD
that there's no need to customer the mapping for username. The default
source property of userPrincipalName is an email address and may in fact
be preferable to the mail property that we were recommending.

userPrincipalName is the username the user logs into AzureAD with, and
apparently the email value can be different in some cases.