BUILDKITE_ARTIFACT_UPLOAD_DESTINATION example path more structured.
We previously recommended (by way of example) `/$BUILDKITE_JOB_ID` as
the prefix within the customer-hosted artifact bucket, which means a
top-level prefix is created for each job, with the artifact files
therein.

For Buildkite-hosted storage, we nest jobs inside builds inside
pipelines inside organizations. This makes the bucket far easier to
navigate, which becomes important when (a) there's millions of objects,
and (b) everything is an opaque UUID.

This commit changes the documentation to recommend a similar hierarchy
to the one we use on our end, but without the top-level Organization,
which is omitted mostly because that environment variable doesn't exist
in jobs, so can't be used by a hook.

As a result, customers will be able to use Pipeline and Build as means
to navigate and maintain their bucket, e.g. setting pipeline-specific
lifecycle policies.