A few tweeks to the SSO docs for Google Workspace (SAML)
We recently re-discovered that Buildkite is listed in the preconfigured
app catalog in Google Workspace, so I've updated the docs to recommend
that path.

It doesn't simplify the process like I hoped it would, but one day we
hope to convince google to enable SCIM on our app so it'll be good to
encourage folks to use it.

While testing the revised process out, I also discovered that it can sometimes
take many hours for changes top the SAML app to take effect, and while
you wait the errors are confusing. I've added a notice to warn future
travellers.

Finally, I added a short note at the top that describes how user can
login to Buildkite once the process is complete. It's kind of neat - a
Buildkite icon appears in the same menu as gmail/calendar/youtube/etc.