Added troubleshooting notes about wait and block steps
Added troubleshooting notes about wait and block steps when they are placed one after another.

Issue context: When a wait step is surrounded by a block step, the wait step is ignored. This has been previously reported by our customers a few times. So updating in our docs, so it is obvious that it's a limitation.

Here is the most recent observation from our customer: https://secure.helpscout.net/conversation/1696682218/27449/