Updating Env Var anchors to be parsed correctly
Nokigiri wasn't parsing the links correctly in our tests, but that
turns out to be because one of our anchors included an asterisk.

That shouldn't be invalid HTML5 but it is invalid HTML4 and seems like
a bad idea when using CSS selectors as our way of navigating the nodes.

This commit subs out the asterisk. A more complete solution is probably
possible, but seems unwarranted for now. Famous last words?