Refactor Test Analytics collector environment variables docs
This moves the test collector environment variable coverage out of the
Ruby Collectors page, into a standalone reference about what kind of
information test collectors read from the environment and upload to
Buildkite.

In this framing, the Buildkite collectors do ✨magic✨ and this doc
discloses what is or is not magical, but not how the magic works (to
transform certain environment variables into URLs, for instance).