Fix spacing of buildkite-agent env
The whitespace between the `Usage` header and the syntax for the command
command is not the same for `env` as it is for other commands as of
v3.39.1. The doc generator is too fragile to tolerate this and does not
output the correct markdown syntax to render the command syntax in a
fixed width font.