Allow muffet to sleep for a bit so that rails can start
We need to wait until rails has started before running muffet as otherwise it will error out
and the test will appear to have failed without having run. The time to wait is hard to
predict, and furthermore, some paths take longer to be be ready than others. The path in this
loop was chosen after some non-systematic observations. So it does not guarantee that the
server will be ready. But it seems to work well in practice.