Fix Rspec link to instance_of
Relishapp has historically been an official source of ruby documentation. However, the maintainer
was recently made redundant from his job and the website is currently returning a 503. The ruby
community will probably either fix the website or find a new site to host this content; however, in
the mean time we're linking to the Internet Archive as a stop gap.

Co-authored-by: Amy Liu <amy.l@buildkite.com>
Co-authored-by: Steven Webb <steve@buildkite.com>