feat: add documentation for new usage page
we're about to launch a new usage page for job minutes and test executions.
this PR adds the documentation for the new page.

There isn't an overarching section for information on both Job
Minutes and Test Analytics, so I've duplicated thigns a bit across
each section.

The job minutes / usage page has been updated, and a new test analytics
page added which is title 'Test Executions'. Could be convinced that
this should go somewhere else, but seemed logical enough for a start.