Clarification on pipeline upload order
When using multiple dynamic pipelines from the same step, these are shown in reverse order in the UI (i.e. steps uploaded in the second call to pipeline upload show up before the first call). This is confusing for some people and there's a workaround if they want to see them in the "expected" order