Update nav.yml
This is a proposed nav update for @mbelton-buildkite to have a look at.

It adds an evaluating buildkite page which adds like a directory of information. 

It also adds getting started guides according to your organization size, infrastructure preferences and code architecture. 

Note this is only a thought exercise, I am not planning to merge it until those pages are incrementally added.