Update dependencies.md
This PR adds information regarding behavior we can expect when a job is cancelled with respective to allow_dependency_failure set to true.

For example if a pipeline has two steps A and B where B has depends_on: A but has allow_dependency_failure set to true so that in case of failure of step A, step B can still run. However when step A is cancelled then in that scenario even if allow_dependency_failure is set to true step B will not run.