Update versions of software on the stack
The version number that have been removed are horribly out of date. It
is a large burden to update them all, so we are removing them.

We will investigate a better way to keep these up to date in the future,
but in the meantime, we will remove them as they are misleading.