Remove image dimensions from SVG diagram images
Ben and I took a look at the image rendering helpers in docs, docs
[renders the image in a responsive image container][1] if a width and
height are specified, which [includes a css class][2] that applies
border and padding.

For svg diagrams, we don’t want to include this treatment and display
the image full width. After briefly considering a new helper, we
decided for now to just omit width and height.

[1]: https://github.com/buildkite/docs/blob/2027f6fcc3db928ff9edb7fd6e8c1b1cbf8c531e/app/models/page.rb#L44-L53
[2]: https://github.com/buildkite/docs/blob/2027f6fcc3db928ff9edb7fd6e8c1b1cbf8c531e/app/models/page.rb#L95

Co-authored-by: Ben Christensen <ben@buildkite.com>