Visual tweaks post layout frontmatter change
Let’s not specify max-width on typographic elements; the page container
has the max prose width on it when a table of contents is present, that
way, typographic elements will simply be as wide as their container
allows them to be.

Note that I’ve tried a pretty brittle way of styling the max width of
pages without a table of contents; namely, on large screens if the
`Page__content` is the second child (the TOC is currently rendered in
markup prior to the main content), it is assumed there is no table of
contents and the page gets the more generous 896px layout.