Upgrade from debian 11 (bullseye) to debian 12 (bookworm)
Debian 12 was released in June 2023 and has a range of package updates.
The vast majority will not be significant in the context of the docs app
and should have no impact on the app behaviour.

Updating is useful though, mainly just so we don't fall behind. It will
also avoid some CVEs being detected by scanning software.

[1] https://wiki.debian.org/DebianBookworm