Add detailed instructions for plugin git repositories
When developing a plugin, we can either use a public git repository
or a private one. We can also use any git host.

Clarify how this is possible by giving an absolute URL and the
required format.

Thanks to Tom Watt for the pointer, this commit should help others
developing plugins and hitting similar pitfalls.