Move Content Security Policy reports to report-uri.com
We've been sending CSP reports to a route on our own site for years, but
never looking at them or taking action.

We're trialling report-uri.com as a cheap hosted option for collecting
these reports and (hopefully) discovering actionable improvements we can
make to our Javascript security [1]

So far we're not thrilled with report-uri.com, but also we're not
sending it all our buildkite.com data. Lets throw it all at them, and
see what the data looks like.

[1] https://3.basecamp.com/3453178/buckets/34063888/messages/6981566345