Added docs for Helm OCI registries
In addition to OCI registries, we will soon be adding support for Helm
native registries. These are a separate registry type.  I'm not sure if
it will be better to have them documented on the same page or separate
pages. For the time being I decided to use one page; therefore, I named
this Helm rather than Helm OCI.