Add standard eventbridge fields to the example payloads
The example payloads for eventbridge events only showed fields that buildkite varies. Copy and
pasting this into the eventbridge sandbox as-is results in an error message that the sample
event is not valid.
Putting the full payload in the examples will make it easier for users to test patterns using the
eventbridge sandbox.