Fix the daily graphql pipeline
I have changed this pipeline to clone over HTTPS, our preferred protocol
in CI. However, that means the default origin in the cloned repo is
readonly.

We have a GH token available in the environment for API calls, so we can
use that to push a branch directly via HTTPS. No ened for an origin on
the repo.

I also found the function to open a PR requires a 3rd arg now, and a
recent commit dropped the colon from the auth header.