update netlify-cli package
While working in this repo last week, I noticed that the version of
netlify-cli we were using was ~18 months old. JS dependencies that
old make me itch - it's far too easy to fall behind enough that version
incompatibilities in various transitive dependencies make it hard to get
back up to date.

It seems like this is a CI tool that's not used in production, and our
usage of it is pretty light. Maybe bumping it to the most recent version
will Just Work?