Serve CSP headers without a report-uri in non-production environments
We're serving CSP headers like this in all environments:

    content-security-policy-report-only: object-src 'none'; ....; report-uri https://buildkite.uriports.com/reports/report

That means we're reporting lots of CSP violations to uriports in dev and
test, which is consuming a significant portion of our quota. This change
will mean we continue to serve CSP headers on responses, but we can
leave CSP_REPORT_URI unset outside of production.

I could've also used Rails.env.production? however, I prefer to avoid
switching on environment booleans like that. Checking the contents of
ENV lets us vary the behaviour in each environment without a code
change.