Update netlify-cli to the latest 17.x release
This includes a bump to path-to-regexp, a transitive that GH has flagged
as a ReDOC "security alert".

We don't care about ReDOS, but we may as well use it as an excuse to
bump a few versions.

There is a brand new 18.x release of netlify-cli, but I haven't tried
it. Maybe next time.