Switch autoloading to using zeitwerk
The docs site currently runs rails 6.1, which is preettty old. 8.0 is
the latest. It'd be nice to get to 7.x at least, to avoid falling too
far behind.

I plan to open a PR that upgrades to rails 7.0, but following the rails
upgrade guide [1] and zeitwork migration guide [2], I first need to
change the autoloading mode from `classic` to `zeitwerk`. Zeitwork is
the only autoloading option in 7.0, so we gotta use it.

Fortunately, this is a super simple app and the switch is pretty
painless. I had to tweak a single constant to make it fit zeitwerks
requirement that the class name be a CamelCase version of the file name

[1] https://guides.rubyonrails.org/v7.0/upgrading_ruby_on_rails.html
[2] https://guides.rubyonrails.org/v7.0/classic_to_zeitwerk_howto.html