Upgrade to rails 7.0.x
Once the prep in #230 has merged, we can do the actual upgrade.

Thanks to the docs app being so simple (no database!) this most Just
Works. The only tricky part was needing to formally register a mime type
for markdown, and map it to the `:md` symbol for rails.

Mostly we manage rendering all the markdown in `pages/` with custom
code, however our shared partial code (app/models/page.rb:74) uses:

    ActionController::Base.render(partial: '...', formats: [:md, :html])

In rails 7.0 it seems the `md` in formats is ignored unless the mime
type is registerd.

We can remove the explicit zeitwerk setting now, it's the default in
rails 7.0.

Also resolves a low secerity Github Security alert that doesn't impact us,
but is nice to make formally irrelevant [1] [2]

[1] https://github.com/buildkite/docs-private/security/dependabot/26
[2] https://github.com/buildkite/docs-private/pull/160