Upgrade to rails 7.1
We're running rails 7.0, firest released in 2021. This gets us to rails
7.1, first released in 2023.

The app is so simple (no database!), this seems to Just Work.

The Gemfile.lock changes are noisy, but it seems to mostly be adding
explicit dependencies on stdlib libraries that have been extracte to
gems. No doubt this will help when we try upgrading to ruby 3.4