Upgrade to rails 7.2
We're on rails 7.1, first released in 2023. This upgrades us to the
rails 7.3 series, first released in 2024.

The app is so simple (no database!) this seems to Just Work.

A small change to a feature test was required. An exception is no longer
raised in the Test env, but we can assert that the expected content
(from PagesController) are rendered to the page