Upgrade to ruby 3.4
We get a bundler update as part of that - ruby 3.4 ships with bundler
2.6. This seems to Just Work.

graphql gets a bump for compatibility with 3.4. The graphql gem is
famously unstable and I normally wouldn't YOLO update it like this.
However in this project we only use it in dev and CI, never in prod. I
can see that the CI steps that check the graphql schema are passing, so
I think it's fine. If there's any surprises, there shouldn't be any
customer impact and we can fix in a follow up.

I included small updates to foreman and stringex as well - they were
generating warnings about needing an explicit dependency on `ostruct`
before ruby 3.5 is released.