Update netlify-cli to 18.x
This will resolve two `moderate` GH security alerts. Neither is
concerning, but we may as well sort them out while I'm in
update-all-the-things mode.

The release notes for 18.0.0 list two breaking changes, but neither
looks like it'll impact us. We'll find out soon enough if the netlify
preview for this PR fails to create.