Document the `--aws-session-tag` flag to `buildkite-agent oidc request-token`
This flag was added in agent version v3.83.0 of the agent, and is useful
for the common use case of assuming an AWS IAM role using OIDC.

In [1] I've confirmed how the flag works and built support for it into
our official plugin for AWS+OIDC, but I thought it was worth leaving a
few breadcrumbs of documentation about the low level flag and command as
well. No doubt some customer will find it helpful.

I'd like to also expand the higher level documentation on using OIDC to
integrate with AWS [2], but I'll tackle that in a separate PR.

[1] https://github.com/buildkite-plugins/aws-assume-role-with-web-identity-buildkite-plugin/pull/18
[2] https://buildkite.com/docs/pipelines/security/oidc/aws