Make GitHub webhook setup instructions more explicit
We have customers creating webhooks with the default content-type of
`application/x-www-form-urlencoded` (which will error if the webhook
payload is larger than 4.15MB), and we've also seen webhooks
delivering events that we will never process.

Updating the docs to be more specific about what we expect the
configuration of a webhook to be.