chore: update the cluster api documentation to include maintainers
As a part of work on the mcp server I have run into some issues with handling of maintainers in the list, get and create cluster api, this highlighted a gap in our documentation. I have updated the clusters section to include example output and fixed a couple of typos.